requirements/base.txt
Django==5.1
django-split-settings==1.2.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.8
requests==2.32.3
django-minio-backend==3.8.0
Pillow==11.1.0
django-filter==24.3
